move logic view up into measure view, remove allPatients references
The logic view is not going to be contained within the calculation view. Instead of binding them by composition, they're now bound by events. The logic view registers for these events within the measure view, when the logic view is instantiated. (This might change if the calculation view is recreated upon selecting a different population.)

The measure calculation view was also being given the measure's patients within an `allPatients` variable; the measure's patients should be exposed to the calculation view's context already via the `patients` property, so this is a pretty clean refactoring.