More style changes to the edit data criteria dialog
- Start and Stop times both have separate date and time fields and they
  fit correctly
- Added updated design for Value and Field Value field
    - Static right now, but format is more or less correct
    - FV is missing the dropdown

Note: Changed spacing between dropdown menu and first input because the
type of inputs change based on the selected value. Needed a separate div
for what changes. Possible to remove spacing, but it wouldn't make a lot
of sense when looking at it with a select instead of text inputs.