protect from measures that can be removed. also, spec reconfig
Some defensive programming to support measures that may be removed from the system. There's also a start on work to reconfigure the specs to only use source data criteria that belong to measures that are also within the test system, though there's more work to be done here. (THIS COMMIT BREAKS THE SPECS.)