Updating to Rails 4. HDS simplexml_parser bonnie_bundler and hqmf2js dependencies have been updated to point to their respective master branches which have been updated to rails 4.
Updated jasmine to the latest version to support rails 4.  This required
updating some of the jamsime test suite.  the waitFor and run functions
have been removed from jamsine and the spyOn utility has changed its
functionality somewhat.

The registrations controller has been updated with a new method to set
the crosswalk_enabled attribute on the user model.  In rails 4 attribute
protection is now the concern of the controller.  Devise has been updated
to align with this methodology so the controller needed to be made aware
of the crosswalk_enabled attribute and allow it to pass through.