Fix for highlighting
Added a new type of VAR to variable_logic_view.js.coffee.
In highlightPatientData (population_logic_view.js.coffee) add logic to
let VARiables bypass the test for being in the finalSpecifics.
In data_criteria_logic_view.js.coffee added logic to check for
Thorax.Views.VariablesLogic so that it will get the right type for
VARiables.
In data_criteria.hbs removed the if condition so that highlight-target
gets added to derived elements.