Fixes for updating actual_values
Found that when a patient is in more than one measure the position in
the actual_values array does not reflect the population_index
accurately.  Had to find a way to update the right population_index for
the right measure.
In the rake found that it was easier to just clear everything (which
there shouldn’t be anything there) and start from a clean slate.