Update the code so that the calculation results are stored in the database
These changes are for two reasons.  The first is to speed the initial
loading of Bonnie by having the launch page populated with cached
results instead of having to calculate the results for every patient.
The second reason is to provide more detail for measure and patient
updates by providing a way to see what has changed in terms of the
rationale for each patient.