Found another case where the Bonnie text diverged from the MAT.
I observed in measure_visual.js.coffee this line:
        parseValue = (value) -> "(result#{getValue(value)})"
It uses lowercase, which matches the capitalization in the MAT html for the population criteria section.

Adding the "Result: " text is trivial in this requested place, but check out CMS53V5. The "attribute value sets" looks good, but the existing logic the splits based on ":" makes the overlapping value sets lose the first half (i.e. the text after this code change is unchanged).