Tweaks to data criteria display (#672)
* Tweaks to data criteria display

- added negation rationale to data criteria
- revised attribute criteria to use value set display name

* Addressing pull request comment about silly oversight in description of value set

* variation of text, now data criteria of negation rationale matches logic display

* Made attribute value sets text match human readable data criteria from MAT

- also split null check into separate line to remove ambiguity

* Found another case where the Bonnie text diverged from the MAT.

I observed in measure_visual.js.coffee this line:
        parseValue = (value) -> "(result#{getValue(value)})"
It uses lowercase, which matches the capitalization in the MAT html for the population criteria section.

Adding the "Result: " text is trivial in this requested place, but check out CMS53V5. The "attribute value sets" looks good, but the existing logic the splits based on ":" makes the overlapping value sets lose the first half (i.e. the text after this code change is unchanged).