Made Bonnie Logo unclickable when navbar present.
Fix for:
"Redundant Links: In the header the Dashboard and Bonnie logos go to the same
place. Bonnie Logo should not be clickable."

JIRA Ticket: https://jira.mitre.org/browse/BONNIE-1016

NOTE: There are several pages without the navbar that still have the Bonnie logo
as a link to the main page.  The WAVE Tool did not complain about those pages,
so I left them unaltered.