Cql4bonnie 508 fixes (#849)
* Remove "justify" text-align for 508 compliance.

* Remove double label for Deceased

* Make the text in the breadcrumbs darker

* Fix contrast issues on patient builder page
- Make the elements panels more contrasty
- Make the grey for selected codes darker (similiar to the breadcrumbs)

* Tweak the highlight colors for data elements.

* Made Bonnie Logo unclickable when navbar present.

Fix for:
"Redundant Links: In the header the Dashboard and Bonnie logos go to the same
place. Bonnie Logo should not be clickable."

JIRA Ticket: https://jira.mitre.org/browse/BONNIE-1016

NOTE: There are several pages without the navbar that still have the Bonnie logo
as a link to the main page.  The WAVE Tool did not complain about those pages,
so I left them unaltered.

* Changes to Account Edit page for 508 compliance

1. Renamed label that had 'for' field duplicated with another input to address:
https://jira.mitre.org/browse/BONNIE-1023

2. Added id to input to match label to address
https://jira.mitre.org/browse/BONNIE-1024

3. Fix minor typo

* Made Measure List Page conform to 508 requirements

Added "for" field to percentage label and matching "id" field in the input to
link the two to address:
https://jira.mitre.org/browse/BONNIE-1019

* Connect coverage label with input

* Address duplicated label on Measure Dashboard

Each measure needs a unique ID so that there are no duplicated label/input pairs
for 508 compliance.  This change generates a random pseudo-guid to append to the
  id to ensure they are all different.

* Make the text for Dashboard, Account, Admin darker when selected.
https://jira.mitre.org/browse/BONNIE-1015

* Remove comments (that were being used to hide the extra label).
https://jira.mitre.org/browse/BONNIE-1020

* Increase contrast: Measure Title, CMSID, Loading

* Update colors to increase constrast for 508

Made new color variable @gray-darkest for #494949

Changed red color for "btn-danger"

Made caption gray color darker (there is only one caption in Bonnie, and it is
hidden anyway, but the tool was throwing an error).

* Address PR Comments

1. Remove duplicated code for guid-generator function

2. Change landing page links in footer to be darker blue for 508 contrast.

* switch two dark blues to use css variable instead