Changes medication units to be ucum compliant
Used 2 references for what these units should be:
1) http://hl7.org/fhir/us/core/2017Jan/ValueSet-us-core-ucum.html
   - {Capsule} for capsule(s)
   - {Dose} for dose(s)

2) http://unitsofmeasure.org/ucum.html#section-Examples-for-some-Non-Units.
   - g for gram(s)
   - ug for mcg(s)
   - mL for ml(s)
   - {tbl} for tablet(s)

Added 2 rake tasks, intended to be run just once:
- bonnie:patients:fix_non_ucum_dose_and_quantity_dispensed
- bonnie:patients:tabulate_dosage_units

The tabulate task is to count intances, and the fix task does the conversion.
Thereforeone one could run the tabulate, the fix, and the tabulate a second
time to see the changed results.