Valuesetsbyoid cache (#899)
* cache refactoring of bonnie.valueSetsByOid in valueSetsForCodeService()

* Remove outdated jasmine test.

In writing the unit test for this PR, I discovered a syntax mistake that existed
in some other tests, including the one removed with this commit.

The mistake is not using () in a Jasmine assertion:
`expect(true).toBeFalse` will return undefined, which passes the test.
`expect(true).toBeFalse()` is the correct syntax.

When I updated the removed test to the correct syntax, it failed.  Further
investigation showed that the test is outdated because this view does not
contain the elements it is expecting.