Fix reporting bug in update_source_data_criteria
Also discovered a problem with using 'title' to update the
source_data_criteria.  There can be multiple data criteria on a measure
with the same title and description.

Removing this change causes 56 patients to be unable to be updated due
to mismatching ids.

We may need to find a way to map these paients to data_criteria on a
measure or come up with a process for otherwise updaing them.