When building the codeSystemMap in the measure model, parse unknown code system names from their uri.
When building the contents for the value set / drc dropdowns, the logic  defers to the codeSystemMap before attempting to parse a human readable name from the uri. However, the codeSystemMap already generates entries for missing code systems. As a result, the logic in the input view to parse the human readable from the uri is never reached. Moving the name parsing logic to the measure model.