Removing reference attributes from sibling resources that target the resource being deleted.
When removing a Reference attribute, do not delete the target resource in case it is used elsewhere in the patient history. For example, referenced by another resource or was meanginfully added prior to be referenced. Target resource is being replaced when the attribute is overwritten, per our usual attribute modification pattern.

Displaying FHIR ID on Resource Card to help users match Reference attributes with their target Resource.

Adding Condition Resource and a Reference attribute to that condition in unit test fixture.

Adding test case for removing reference attribute when targeted resource is deleted.

Splitting out the criteria spec to reduce likelihood of execution timeout.

Grabbing the FHIR ID from the dataElement for display on the resource card.